+++ api/database_engine.go
@@ -17,8 +17,6 @@
 package api
 
 import (
-	"context"
-	"errors"
 	"fmt"
 	"net/http"
 	"strings"
@@ -96,12 +94,6 @@ func (e *EverestServer) UpgradeDatabaseEngineOperator(ctx echo.Context, namespac
 	return nil
 }
 
-type upgradePreflightCheckArgs struct {
-	targetVersion  string
-	engine         *everestv1alpha1.DatabaseEngine
-	versionService versionservice.Interface
-}
-
 // GetOperatorUpgradePreflight gets the preflight check results for upgrading the specified database engine operator.
 func (e *EverestServer) GetOperatorUpgradePreflight(
 	ctx echo.Context,
@@ -150,83 +142,4 @@ func validateDatabaseEngineOperatorUpgradeParams(req *DatabaseEngineOperatorUpgr
 		return fmt.Errorf("invalid target version '%s'", req.TargetVersion)
 	}
 	return nil
-}
-
-func (e *EverestServer) runOperatorUpgradePreflightChecks(
-	ctx context.Context,
-	dbs []everestv1alpha1.DatabaseCluster,
-	args upgradePreflightCheckArgs,
-) (*OperatorUpgradePreflight, error) {
-	dbResults := make([]OperatorUpgradePreflightForDatabase, 0, len(dbs))
-	for _, db := range dbs {
-		// Check that the database engine is at the desired version.
-		engineVersionValid, reqEngineVersion, err := e.validateDatabaseEngineVersionForOperatorUpgrade(ctx, db, args)
-		if err != nil {
-			return nil, errors.Join(err, errors.New("failed to validate database engine version for operator upgrade"))
-		}
-		if !engineVersionValid {
-			dbResults = append(dbResults, OperatorUpgradePreflightForDatabase{
-				Name:        pointer.To(db.GetName()),
-				PendingTask: pointer.To(UpgradeEngine),
-				Message: pointer.ToString(
-					fmt.Sprintf("Upgrade DB version to %s", reqEngineVersion)),
-			})
-			continue
-		}
-
-		// Check that DB is at recommended CRVersion.
-		if recCRVersion := db.Status.RecommendedCRVersion; recCRVersion != nil {
-			dbResults = append(dbResults, OperatorUpgradePreflightForDatabase{
-				Name:        pointer.To(db.GetName()),
-				PendingTask: pointer.To(Restart),
-				Message: pointer.ToString(
-					fmt.Sprintf("Update CRVersion to %s", *recCRVersion)),
-			})
-			continue
-		}
-
-		// Database is in desired state for performing operator upgrade.
-		dbResults = append(dbResults, OperatorUpgradePreflightForDatabase{
-			Name:        pointer.To(db.GetName()),
-			PendingTask: pointer.To(Ready),
-		})
-	}
-	return &OperatorUpgradePreflight{
-		Databases:      &dbResults,
-		CurrentVersion: pointer.To(args.engine.Status.OperatorVersion),
-	}, nil
-}
-
-// validateDatabaseEngineVersionForOperatorUpgrade validates that the provided database cluster
-// is at the desired version to upgrade the operator to the targetVersion.
-func (e *EverestServer) validateDatabaseEngineVersionForOperatorUpgrade(
-	ctx context.Context,
-	database everestv1alpha1.DatabaseCluster,
-	args upgradePreflightCheckArgs,
-) (bool, string, error) {
-	engineType := args.engine.Spec.Type
-	operator, found := versionservice.EngineTypeToOperatorName[engineType]
-	if !found {
-		return false, "", fmt.Errorf("unsupported engine type %s", engineType)
-	}
-
-	supportedVersions, err := args.versionService.GetSupportedEngineVersions(ctx, operator, args.targetVersion)
-	if err != nil {
-		return false, "", err
-	}
-
-	// GetSupportedEngineVersions always returns a non-zero length result.
-	minVersion := supportedVersions[0]
-
-	if !strings.HasPrefix(minVersion, "v") {
-		minVersion = "v" + minVersion
-	}
-	currentVersion := database.Spec.Engine.Version
-	if !strings.HasPrefix(currentVersion, "v") {
-		currentVersion = "v" + currentVersion
-	}
-
-	// Current engine version should be greater than or equal to the minimum supported version.
-	versionValid := semver.Compare(currentVersion, minVersion) >= 0
-	return versionValid, strings.TrimPrefix(minVersion, "v"), nil
 }
